django-projects-2024/Github-Data-Mining/pyproject.toml
[tool.poetry]
name = "github-data-mining"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygithub = "^2.5.0"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
